Add new classes for project reconfiguration
While making the renderer work well, I realized to need to make a class to manage any class type. Because I want to test the ECS code that requires a system that get components as the component type.
That classes is called "Class" and "ClassType" added by me. And also I made class called "EntityInterface" that can manage and get components and entities.
Besides that, I worked more. Overall, the project structure was rebuilt. But we won't be able to see any models or materials in the scene. although I rebuilt it, I couldn't work on root signatures, pipeline objects, etc.